Resolved unicode issue while printing title
When the title was printed using the print. It was showing the following error:
'ascii' codec can't encode character u'\u2013' in position 9: ordinal not in range(128)

When have isses with 2013 appearing in one of the text: "A Buddhist monk confronts Japan?s suicide culture (2013) "
its taking 2013 as unicode and goes into this error.
We encode the string in unicode and it solves the isse.